separate the concepts of tweaking vs removing
eg: we can "strip" comment lines without removing them,
or we can optionally remove them entirely.

this also changes tag stripping to only strip lines which contain
nothing but tags, which is the actual use-case I need it for -
and neatly bypasses all edge cases. And of course these lines
can either be "cleared" or entirely removed.

as part of this, newlines are preserved by default, even on comments
which span multiple lines.