major refactor: use remark markdown parser
rather than using a naive, mostly line-based parser to strip out
comments and tag-lines, switch to using remark (which Obsidian itself
uses internally). This gets rid of almost all known edge-cases such as
misdetecting %%comments%% and #tags within <html attributes=""> and
`inline-code literals`. Known remaining edge-cases are now noted in the
README file.

This had many implications, including:

 - switching to jest, which has more support for different dialects of
   javascript (via Babel).
 - creation of many (though not complete) unit tests, to ensure
   edge-cases are handled (or in most cases, completely avoided).
 - refactoring the way various aspects of the sections are detected,
   mostly with the aim of making testing easier/narrower.
 - making use of the existing section/heading cache, which should make
   "headed section detection" faster and more accurate.
 - splitting the entire plugin into a lot more, smaller files, to make
   everything easier to reason about.

And of course, all of it was an unreadable mess which went through many
internal iterations, the majority of which were completely broken in
numerous and interesting ways. So this commit is a squash of all that
work, hopefully resulting in a basis that any future work can make much
smaller/incremental changes to.