fix: "copy file" only copies first section
I'm not sure how this made it through manual testing (though, there
still aren't any automated tests for this), but: "copy all" was only
copying the first section.

Not only did this mean that anything after the first header was being
skipped, it also meant that anything *before* the first header, notably
the metadata and any un-headed text, was also being skipped.