do not abort when headings are unknown
previously, we checked whether the "headings" had been processed in
order to determine if the file was successfully loaded from cache.

when there are NO headings, however, this was left unset, resulting in
an intentionally thrown (and in practice caught) error. sectionMarkdown
already handles the "no headers" situation with sane fallbacks (eg: not
iterating over an empty list), so the fix is trivial: just don't check
whether or not there are any headers before we start processing the
markdown.